+++ src/terminal/hardio.h
@@ -1,9 +1,9 @@
-/* $Id: hardio.h,v 1.1 2003/05/04 17:51:07 pasky Exp $ */
+/* $Id: hardio.h,v 1.2 2005/04/27 15:15:01 jonas Exp $ */
 
 #ifndef EL__TERMINAL_HARDIO_H
 #define EL__TERMINAL_HARDIO_H
 
-int hard_write(int, unsigned char *, int);
-int hard_read(int, unsigned char *, int);
+ssize_t hard_write(int, unsigned char *, int);
+ssize_t hard_read(int, unsigned char *, int);
 
 #endif
